fix(store): pinTab idempotent guard + updateTab strips pinned/locked
Review fixes:
- pinTab on already-pinned tab is now no-op (preserves position)
- updateTab strips pinned/locked fields to protect invariants
- Split useTabStore tests into base + pin-lock files (both <200 lines)
